clean up api::discovery module - for #523
- moves the modules from kube::client to kube::api
- introduces easier way to get recommended_resources and
  recommended_kind from an `ApiGroup`
- restructures file for readability (internal Version sorting quirks
  documented and tested)
- renamed `ApiResourceExtras` to `ApiCapabilities`
- renamed `ApiGroup::group` to `ApiGroup::get`
- added support for `Discovery::single` which avoids having to deal with
  the `HashMap` and returns a straight `ApiGroup` (making the
recommended doc examples more compelling)
- documents the version sorting a bit better
- renamed `ApiGroup::preferred_version_or_guess` to
  `ApiGroup::preferred_version_or_latest` and referred to the version
policy
- separated querying to helpers on `ApiGroup`
- moved sorting to `ApiGroup` construction part
